% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blkrandomization.R
\name{blkrandomization}
\alias{blkrandomization}
\title{Blocked randomization}
\usage{
blkrandomization(n, blocksize, block)
}
\arguments{
\item{n}{numeric number of subjects who will be randomized}

\item{blocksize}{numeric value of block size used for blocked randomization}

\item{block}{vector of treatment blocks used for blocked randomization}
}
\value{
trt a sequence of treatment assignments
}
\description{
The fuction is used to generate treatment assignments based on blocked randomization.
}
\examples{
blocksize <- 4
block <- c(1, 2, 3, 4) # treatment 1, 2, 3, 4
n <- 35
blkrandomization(n, blocksize, block)
}
