% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randbalance.R
\name{randbalance}
\alias{randbalance}
\title{Displays the number of randomized subjects at each level for all covariate factors.}
\usage{
randbalance(trt, covmat, ntrt, trtseq)
}
\arguments{
\item{trt}{treatment sequence for all the randomized subjects}

\item{covmat}{matrix or data frame of covariate factors}

\item{ntrt}{numeric number of treatment groups}

\item{trtseq}{vector of a sequence of treatment groups}
}
\value{
the number of randomized subjects at each level for all covariate factors
}
\description{
The fuction to cound the number of randomized subjects at each level for all covariate factors
}
