% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totimbal.R
\name{totimbal}
\alias{totimbal}
\title{Calculates the total imbalance measured by minimization algorithms.}
\usage{
totimbal(trt = trt, covmat = covmat, covwt = covwt, ratio = ratio,
  ntrt = ntrt, trtseq = trtseq, method = "Range")
}
\arguments{
\item{trt}{treatment sequence for all the randomized subjects}

\item{covmat}{matrix or data frame of covariate factors}

\item{covwt}{vector of weights of the covaraite factors}

\item{ratio}{vector of randomization ratios for each treatment}

\item{ntrt}{numeric number of treatment groups}

\item{trtseq}{vector of a sequence of treatment groups}

\item{method}{the method or algorithm for the minimization randomization}
}
\value{
total imbalance
}
\description{
The function to calculates the total imbalance measured by minimization algorithms
}
